Modularlized electron.js into separate files. See api-components/ to edit file behavior and redirectAuthenticate function
